requirements.txt
django_environ==0.11.2
psutil==5.9.5
python-telegram-bot==20.6
ipython==8.16.1
requests==2.31.0
urllib3==2.0.6
